// kiar-ui/openapi/model/transformerConfig.ts
/**
 * KIAR Dashboard API
 * API for the KIAR Dashboard.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TransformerType } from './transformerType';


export interface TransformerConfig { 
    type: TransformerType;
    parameters: { [key: string]: string; };
}
export namespace TransformerConfig {
}